Add callsign search route, refactor queries into module
- Add module for search
    - Add search by callsign
- Refactor QSL table to be a component and make it slightly more flexible
- Refactor queries.py into a module
- Fix gitignore to not ignore new files under app/database
