+++ app/blueprints/search/__init__.py
@@ -1,0 +1,2 @@
+from .base import bp
+from .callsign import callsign
