+++ app/blueprints/search/base.py
@@ -1,0 +1,3 @@
+from flask.blueprints import Blueprint
+
+bp = Blueprint(name="search", import_name=__name__)
