+++ app/database/queries/__init__.py
@@ -1,0 +1,13 @@
+from .functional import (
+    ensure_user,
+    get_object,
+    get_user,
+    is_unique_qso,
+    qso_report_exists,
+)
+from .map import (
+    get_user_qsos_for_map_by_rxqso,
+    get_user_qsos_for_map_by_rxqso_count,
+)
+from .qso_page import get_25_most_recent_rxqsls
+from .search import callsign
